refactor: pure tests
