+++ crates/cdk-integration-tests/Cargo.toml
@@ -14,6 +14,7 @@ rust-version = "1.63.0"                                                         
 http_subscription = ["cdk/http_subscription"]
 
 [dependencies]
+async-trait = "0.1"
 axum = "0.6.20"
 rand = "0.8.5"
 bip39 = { version = "2.0", features = ["rand"] }
@@ -57,8 +58,6 @@ getrandom = { version = "0.2", features = ["js"] }
 instant = { version = "0.1", features = ["wasm-bindgen", "inaccurate"] }
 
 [dev-dependencies]
-async-trait = "0.1"
-rand = "0.8.5"
 bip39 = { version = "2.0", features = ["rand"] }
 anyhow = "1"
 cdk = { path = "../cdk", features = ["mint", "wallet"] }
